requirements.txt
Flask==1.0.2
Jinja2==2.10
python-dateutil==2.7.3
Flask-Babel==0.11.2